Project-X/requirements.txt
ipykernel
langchain
python-dotenv
langchain-community
langchain-text-splitters
langchain-openai
chromadb
langchain-chroma
streamlit
langchain-groq
langchain-core
fastapi
uvicorn
langserve[all]
sse-starlette